[GILJOB-82] refactor: 파일 export 구조 변경
