--- src/components/atoms/Badge/index.ts
@@ -1,3 +1,0 @@
-import Badge from './Badge';
-
-export default Badge;
