--- src/components/atoms/Text/index.ts
@@ -1,3 +1,0 @@
-import Text from './Text';
-
-export default Text;
