+++ src/components/atoms/index.ts
@@ -1,0 +1,3 @@
+export { default as Badge } from './Badge';
+export { default as SearchBar } from './SearchBar';
+export { default as Text } from './Text';
