+++ src/components/molecules/Card/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import './style.scss';
-import Badge from '@src/components/atoms/Badge';
-import Text from '@src/components/atoms/Text';
+import { Text, Badge } from '@src/components/atoms';
 
 export interface CardProps {
   step: '입문' | '초급' | '중급' | '고급' | '통달';
