+++ src/components/molecules/index.ts
@@ -1,0 +1,4 @@
+export { default as Card } from './Card';
+export { default as Content } from './Content';
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
