+++ src/pages/Landing/Sub/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Text from '@src/components/atoms/Text';
+import { Text } from '@src/components/atoms';
 import './style.scss';
 
 const Sub: React.FC = () => {
