+++ src/pages/Landing/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Based from '@src/pages/Based';
 import Main from './Main';
 import Contents from './Contents';
 import Count from './Count';
@@ -7,12 +6,12 @@ import Sub from './Sub';
 
 const Landing: React.FC = () => {
   return (
-    <Based hasHeader>
+    <>
       <Main />
       <Contents />
       <Count />
       <Sub />
-    </Based>
+    </>
   );
 };
